+++ src/mastodon_api/timelines/types.rs
@@ -1,12 +1,12 @@
 use serde::Deserialize;
 use uuid::Uuid;
 
-fn default_page_size() -> i64 { 20 }
+fn default_page_size() -> u16 { 20 }
 
 #[derive(Deserialize)]
 pub struct TimelineQueryParams {
     pub max_id: Option<Uuid>,
 
     #[serde(default = "default_page_size")]
-    pub limit: i64,
+    pub limit: u16,
 }
